financial_control/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
django-widget-tweaks==1.4.8
psycopg2-binary==2.8
